Insert new show before pulling it back out
Creating the new show by using the alternate option from the get method
doesn't add the key to the dict. The rest of the method assumes it's
using a value from the dict so it doesn't need to save it back. Maintain
that assumption by inserting the new, empty list for the show first.